feat: 新增根据agentId查询操作系统的接口和查询列出用户所有第三方agent的接口
