+++ src/backend/ci/core/environment/biz-environment/src/main/kotlin/com/tencent/devops/environment/resources/thirdPartyAgent/BuildAgentThirdPartyAgentResourceImpl.kt
@@ -183,9 +183,7 @@ class BuildAgentThirdPartyAgentResourceImpl @Autowired constructor(
         return Result(thirdPartyAgentPipelineService.updatePipelineStatus(projectId, agentId, secretKey, response))
     }
 
-    override fun getOs(userId: String, projectId: String, agentId: String): Result<String> {
-        return Result(thirdPartyAgentService.getOs(userId, projectId, agentId))
-    }
+
 
     private fun checkParam(
         projectId: String,
